Remove unused scaffold files and add real tab controllers
