+++ config.py
@@ -1,0 +1,1 @@
+APP_REGISTRY_PATH = r"Software\MyApp"
