+++ controllers/dq_temp_controller.py
@@ -1,0 +1,6 @@
+class DQTempTabController:
+    def __init__(self, main_window):
+        self.main_window = main_window
+
+    def plot(self):
+        self.main_window.plot_Arr()
